+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- `to_dict()` method and `from_dict()` class method to `Note`.  
+- `to_dict()` method and `from_dict()` class method to `Comment`.
+- `to_dict()` method and `from_dict()` class method to `User`.
+
+### Fixed
+- `Note` can now be imported from package.
+- `Comment` can now be imported from package.
+
 ## [0.3.0] - 2023-03-14
 
 ### Added